requirements.txt
# AI/ML libraries - commented for Vercel deployment
# aiohappyeyeballs==2.6.1
# aiohttp==3.13.5
# aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
# av==17.0.1  # audio/video processing
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cloudinary==1.44.2
colorama==0.4.6
cryptography==46.0.7
# ctranslate2==4.7.1  # ML translation
dj-database-url==3.1.2
Django==6.0.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
# edge-tts==7.2.8  # text-to-speech AI
# faster-whisper==1.2.1  # speech-to-text AI
# ffmpeg-python==0.2.0  # media processing
# filelock==3.29.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
future==1.0.0
google-auth==2.49.2
gunicorn==25.3.0
h11==0.16.0
# hf-xet==1.4.3  # HuggingFace tools
httpcore==1.0.9
httpx==0.28.1
# huggingface_hub==1.12.0  # HuggingFace AI models
idna==3.13
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.4
# onnxruntime==1.25.0  # ML runtime
packaging==26.2
pillow==12.2.0
propcache==0.4.1
protobuf==7.34.1
psycopg2-binary==2.9.12
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.20.0
PyJWT==2.12.1
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
sympy==1.14.0
tabulate==0.10.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
whitenoise==6.12.0
yarl==1.23.0
# ckeditor
django-ckeditor==6.7.3
django-js-asset==3.1.2
